Rename and relocate mixer/test/testenv to pkg/mock
Mixer contains mixer/pkg/adapter/test/env.go exposing a test.NewEnv method, which was
confusing with mixer/test/testenv's testenv.NewEnv method. The testenv package is now
renamed to 'mock' and the type is now called Server, giving us a nice mock.NewServer
signature instead.
