+++ mixer/test/e2e/e2e_report_test.go
@@ -23,11 +23,11 @@ import (
 
 	istio_mixer_v1 "istio.io/api/mixer/v1"
 	pb "istio.io/api/mixer/v1/config/descriptor"
+	testEnv "istio.io/istio/mixer/pkg/mock"
 	"istio.io/istio/mixer/pkg/template"
 	spyAdapter "istio.io/istio/mixer/test/spyAdapter"
 	e2eTmpl "istio.io/istio/mixer/test/template"
 	reportTmpl "istio.io/istio/mixer/test/template/report"
-	testEnv "istio.io/istio/mixer/test/testenv"
 )
 
 const (
@@ -163,14 +163,14 @@ func TestReport(t *testing.T) {
 		}
 
 		adapterInfos, spyAdapters := ConstructAdapterInfos(tt.behaviors)
-		env, err := testEnv.NewEnv(&args, e2eTmpl.SupportedTmplInfo, adapterInfos)
+		env, err := testEnv.NewServer(&args, e2eTmpl.SupportedTmplInfo, adapterInfos)
 		if err != nil {
-			t.Fatalf("fail to create testenv: %v", err)
+			t.Fatalf("fail to create mock: %v", err)
 		}
 
 		defer closeHelper(env)
 
-		client, conn, err := env.CreateMixerClient()
+		client, conn, err := env.CreateClient()
 		if err != nil {
 			t.Fatalf("fail to create client connection: %v", err)
 		}
